% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeBatchOfBenchExp.R
\name{analyzeBatchOfBenchExp}
\alias{analyzeBatchOfBenchExp}
\title{analyse batch of benchmark experiments output}
\usage{
analyzeBatchOfBenchExp(benchmarkResult, tasks)
}
\arguments{
\item{benchmarkResult}{a list of class BenchmarkResult returned in the outputs of analyzeBatchOfBenchExp}

\item{tasks}{a list of the mlr taks that were passed to analyzeBatchOfBenchExp}
}
\value{
a list containing relevant performance indications by learners
}
\author{
Thomas Goossens
}
