% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBenchmark.R
\name{makeBenchmark}
\alias{makeBenchmark}
\title{make a mlr model + model metadata for a specific task using a specific learner}
\usage{
makeBenchmark(task, learner)
}
\arguments{
\item{task}{an object of class mlr::makeRegrTask()}

\item{learner}{an object of class mlr::makeLearner()}
}
\value{
an object of class mlr::train()
}
\author{
Thomas Goossens
}
