% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBmrsAnalysisPlots.R
\name{makeBmrsAnalysisPlots}
\alias{makeBmrsAnalysisPlots}
\title{Plot analyse batch of benchmark experiments output}
\usage{
makeBmrsAnalysisPlots(bmrsResult, bmrsAnalysis)
}
\arguments{
\item{bmrsResult}{an object of class mlr::benchmarkResult}

\item{bmrsAnalysis}{a list of class BenchmarkResult returned in the outputs of makeBmrsAnalysisPLots}
}
\value{
a list of plots
}
\author{
Thomas Goossens
}
