% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeComparison.R
\name{makeComparison}
\alias{makeComparison}
\title{make a comparison between 2 stations}
\usage{
makeComparison(dataset, sensor, sids, dfrom, dto)
}
\arguments{
\item{dataset}{a dataframe containing the records from the 2 stations you want to compare. ::TODO ::Must contain sunset and sunrise !}

\item{sensor}{a character specifying the name of the sensor you want to compare}

\item{sid}{a numeric vector specifying the sids you want to compare}
}
\value{
a list which elements are objects of class mlr::benchmark()
}
\author{
Thomas Goossens
}
