% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDescStats.R
\name{makeDescStats}
\alias{makeDescStats}
\title{make the descriptive stats of a sensor for a station on a specific period of time}
\usage{
makeDescStats(records, sensors, stations = unique(records$sid),
  dfrom = NULL, dto = NULL)
}
\arguments{
\item{records}{a list of hourly dataframes or a single dataframe ::TODO ::Must contain sunset and sunrise !}

\item{stations}{a character vector specifying the sid's of the stations to use}

\item{dfrom}{a character vector}

\item{dto}{a character vector}

\item{sensor}{a character vector specifying the name of the sensors you want to get the descriptive statistics}
}
\value{
a list
}
\author{
Thomas Goossens
}
