% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\title{make a mlr model + model metadata for a specific task using a specific learner}
\usage{
makeModel(task, learner)
}
\arguments{
\item{task}{an object of class mlr::makeRegrTask()}

\item{learner}{an object of class mlr::makeLearner()}
}
\value{
a 2 elements named list : (1) snitch and (2) output. snitch is TRUE if function has provided the expected result. output is a named list which contains :
(1) value : an object of class list made of 3 elements which are of classes "Model", "dataframe", and "list"
(2) condition : a character specifying if the functions has encountered success, warning, error
(3) message : the message relative to the condition
}
\examples{

# create the dataset
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T15:00:00Z",
  sensor = "tsa")

# extract a single hour of the dataset
myDataset = myDataset$output$value

# create the tasks
tasks = purrr::map(dataset, makeTask, target = "tsa")

# extract the required part of the tasks
tasks = tasks \%>\% purrr::modify_depth(1, ~.$"output"$"value"$"task")

# show 1 task
myTask = tasks[[1]]

# create the model
myModel = makeModel(
  task = myTask,
  learner = agrometeorLearners$mulLR_lonLatAlt_NA)

# extract the relevant information
myModel = myModel$output$value

# show the results of the LOOCV of this hour forthe choosen learner
head(model$predictions)
head(model$perfs$iters)

# show the model information
model$trained
}
\author{
Thomas Goossens
}
