% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{makePlots}
\alias{makePlots}
\title{make plots for a sensor of a station for a specific period of time}
\usage{
makePlots(records, combs, dfrom = NULL, dto = NULL, interactive = TRUE)
}
\arguments{
\item{records}{a list of hourly dataframes or a single dataframe ::TODO ::Must contain sunset and sunrise !}

\item{combs}{a named character vector specifying the sid's combinations of the stations you want to put on the same plot}

\item{dfrom}{a character vector}

\item{dto}{a character vector}

\item{sensor}{a character vector specifying the name of the sensors you want to get the descriptive statistics}
}
\value{
a list

records = makeDataset(dfrom = ex_dfrom, dto = ex_dto, sensor = c("tsa", "vvt"), stations = unique(stations.df$sid))
}
\author{
Thomas Goossens
}
