% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStationSensorStats.R
\name{makeStationSensorStats}
\alias{makeStationSensorStats}
\title{make the descriptive stats of a sensor for a station for a specific period of time}
\usage{
makeStationSensorStats(records, sensors, stations = unique(records$sid),
  dfrom = NULL, dto = NULL)
}
\arguments{
\item{records}{a list of hourly dataframes or a single dataframe ::TODO ::Must contain sunset and sunrise !}

\item{stations}{a character vector specifying the sid's of the stations to use}

\item{dfrom}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{dto}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{sensor}{a character vector specifying the name of the sensors you want to get the descriptive statistics}
}
\value{
a list
}
\description{
make the descriptive stats of a sensor for a station for a specific period of time
}
\author{
Thomas Goossens
}
