% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStationsSensorStats.R
\name{makeStationsSensorStats}
\alias{makeStationsSensorStats}
\title{make the descriptive stats of sensors of a set of stations for a specific period of time}
\usage{
makeStationsSensorStats(records, sensors, stations = unique(records$sid),
  dfrom = NULL, dto = NULL)
}
\arguments{
\item{records}{a list of hourly dataframes or a single dataframe}

\item{stations}{a character vector specifying the sid's of the stations to characterize}

\item{dfrom}{a datetime string specifying the dateTime
Must have the form \code{"YYYY-MM-DDTHH:MM:SSZ"}}

\item{dto}{a datetime string specifying the dateTime
Must have the form \code{"YYYY-MM-DDTHH:MM:SSZ"}}

\item{sensor}{a character vector specifying the name of the sensors you want to get the descriptive statistics}
}
\value{
a dataframe
}
\description{
make the descriptive stats of sensors of a set of stations for a specific period of time
}
\author{
Thomas Goossens
}
