% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataset.R, R/makeTasks.R
\name{makeTasks}
\alias{makeTasks}
\alias{makeTasks}
\title{make a dataset of stations records}
\usage{
makeTasks(dataset, staticExpl, dynExpl)

makeTasks(dataset, staticExpl, dynExpl)
}
\arguments{
\item{dataset}{a dataframe containing all the hourly/daily records you want to transform a list of mlr tasks}

\item{staticExpl}{a character vector specifying the static explanatory variables you want to add to the taks.
Any combinations of "X", "Y", altitude", "elevation", "slope", "aspect", "Agricultural_areas", "Artificials_surfaces", "Forest", "Herbaceous_vegetation"}

\item{dynExpl}{a character vector specifying the dynamic explanatory variables you want to add.
Any combinations of inca, ens}

\item{stations}{a character specifying the sid's of the stations to use separated by commas}

\item{dfrom}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{dto}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{sensor}{a character specifying the sensor data you want to spatialize.
One of tsa, hct, hra}

\item{stations}{a character specifying the sid's of the stations to use separated by commas}

\item{dynExpl}{a character vector specifying the dynamic explanatory variables you want to add to the task.
Any combinations of inca, ens}
}
\value{
an object of class dataframe containing the desired records

an object of class list containing elements of class mlr::makeRegrTask()
}
\author{
Thomas Goossens

Thomas Goossens
}
