% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrometAPI.R
\name{typeData}
\alias{typeData}
\title{internal function to type data called from agromet}
\usage{
typeData(meta_and_records.l, table_name = "cleandata")
}
\arguments{
\item{meta_and_records.l}{a list containing agromet records and metadata returned by get_from_agromet_api.fun().}

\item{table_name}{a character specifying the name of the agromet table from which the data where called using get_from_agromet_api.fun()}
}
\value{
a typed dataframe
}
\description{
Prepare the data obtained by get_from_agromet_api.fun(). It types all the character variables to their proper types.
}
