% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSpatialization.R
\name{exportSpatialization}
\alias{exportSpatialization}
\title{export a spatialized object either to csv, json or geojson}
\usage{
exportSpatialization(spatialized, path = getwd(), filename = NULL,
  format = "csv", write = FALSE)
}
\arguments{
\item{spatialized}{a dataframe containing the gridded predicted values}

\item{path}{a character specifying the path where you want your geosonfile to be stored. Default = wd}

\item{filename}{a character specifying the name you want to give to the file.}

\item{format}{a character specifying the type of export format. One of "csv", "json" or "geojson". Default = "csv"}

\item{write}{a boolean specifying if fomatted data must be written to file (TRUE) or printed to console (FALSE)}
}
\value{
a list containing a boolean and a character containing the data encoded in the required format. Default = csv
}
\description{
you can both store the encoded string in a variable and in a file by specifying write = TRUE
}
\author{
Thomas Goossens
}
