% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSpatialization.R
\name{exportSpatialization}
\alias{exportSpatialization}
\title{export a spatialized object either to json or geojson}
\usage{
exportSpatialization(spatialized, path = getwd(), filename, csv = FALSE,
  json = FALSE, geojson = FALSE, write = FALSE)
}
\arguments{
\item{spatialized}{a dataframe containing the gridded predicted values}

\item{path}{a character specifying the path where you want your geosonfile to be stored. Default = wd}

\item{filename}{a character specifying the name you want to give to the file.}

\item{csv}{a boolean specifying if data must be exported as csv. Default = FALSE}

\item{json}{a boolean specifying if data must be exported as json. Default = FALSE}

\item{geojson}{a boolean specifying if data must be exported as geojson. Default = FALSE}

\item{write}{a boolean specifying if fomatted data must be written to file (TRUE) or printed to console (FALSE)}
}
\author{
Thomas Goossens
}
