% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSpatialization.R
\name{exportSpatialization}
\alias{exportSpatialization}
\title{export a spatialized object either to csv, json or geojson}
\usage{
exportSpatialization(spatialized, path = getwd(), filename = NULL,
  format = "csv", write = FALSE)
}
\arguments{
\item{spatialized}{a dataframe containing the gridded predicted values}

\item{path}{a character specifying the path where you want your geosonfile to be stored. Default = wd}

\item{filename}{a character specifying the name you want to give to the file. Default = NULL}

\item{format}{a character specifying the type of export format. One of "csv", "json" or "geojson". Default = "csv"}

\item{write}{a snitchean specifying if formatted data must be written to file (TRUE) or printed to console (FALSE)}
}
\value{
a 2 elements named list : (1) bool and (2) output. snitch is TRUE if function has provided the expected result. output is a named list which contains :
(1) value : a character vector containing the data encoded into the desired exportation format
(2) condition : a character specifying if the functions has encountered success, warning, error
(3) message : the message relative to the condition
}
\description{
you can both store the encoded string in a variable and in a file by specifying write = TRUE
}
\author{
Thomas Goossens
}
