% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBenchmark.R
\name{makeBenchmark}
\alias{makeBenchmark}
\title{make a mlr benchmark experiment for various learners on multiple tasks}
\usage{
makeBenchmark(tasks, learners)
}
\arguments{
\item{tasks}{a list which elements are object of class mlr::makeRegrTask()}

\item{learners}{a list which elements are object of class mlr::makeLearner()}
}
\value{
an object of class mlr::makebenchmark()
}
\author{
Thomas Goossens
}
