% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBenchmark.R
\name{makeBenchmark}
\alias{makeBenchmark}
\title{make a multicore parallelized mlr benchmark experiment for various learners on multiple tasks}
\usage{
makeBenchmark(parallel = TRUE, cpus = 4, tasks, learners,
  measures = list(rmse), keep.pred = FALSE, models = FALSE,
  resamplings = "LOO")
}
\arguments{
\item{cpus}{an integer specifying the number of cpus to use for the benchamrk}

\item{tasks}{a list which elements are object of class mlr::makeRegrTask()}

\item{learners}{a list which elements are object of class mlr::makeLearner()}

\item{measures}{a list of the mlr performance metrics you want to get. Default is list(rmse)}

\item{keep.pred}{a boolean specifying if you want to keep the bmr preds. defaut = FALSE}

\item{models}{a boolean specifying if you want to keep the bmr models. defaut = FALSE}

\item{resampling}{a character specifying the type of mlr's CV. Default = LOO}
}
\value{
a list wihch elements are objects of class mlr::benchmark()
}
\author{
Thomas Goossens
}
