% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBmrsAnalysisPLots.R
\name{makeBmrsAnalysisPLots}
\alias{makeBmrsAnalysisPLots}
\title{Plot analyse batch of benchmark experiments output}
\usage{
makeBmrsAnalysisPLots(bmrsAnalysis)
}
\arguments{
\item{bmrsAnalysis}{a list of class BenchmarkResult returned in the outputs of makeBmrsAnalysisPLots}
}
\value{
a list of plots
}
\author{
Thomas Goossens
}
