% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBmrsPlots.R
\name{makeBmrsPLots}
\alias{makeBmrsPLots}
\title{Plot analyse batch of benchmark experiments output}
\usage{
makeBmrsPLots(bmrsExtraction)
}
\arguments{
\item{bmrsExtraction}{a dataframe}
}
\value{
a list of plots
}
\author{
Thomas Goossens
}
