% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{makeDataset}
\alias{makeDataset}
\title{make a dataframe of stations records}
\usage{
makeDataset(stations = paste0(as.character(stations.df$sid), collapse = ","),
  json = NULL, dfrom = NULL, dto = NULL, sensor = "tsa",
  staticExpl = c("X", "Y", "elevation"), dynExpl = NULL)
}
\arguments{
\item{stations}{a character specifying the sid's of the stations to use separated by commas}

\item{json}{a character specifying the path of a json file constructed from a batch db export.
If NULL the API will be called.}

\item{dfrom}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{dto}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{sensor}{a character specifying the sensor data you want to spatialize.
One of tsa, hct, hra}

\item{staticExpl}{a character vector specifying the static explanatory variables you want to add to the taks.
Any combinations of "X", "Y", altitude", "elevation", "slope", "aspect", "Agricultural_areas", "Artificials_surfaces", "Forest", "Herbaceous_vegetation"}

\item{dynExpl}{a character vector specifying the dynamic explanatory variables you want to add to the task.
Any combinations of inca, ens}
}
\value{
an object of class dataframe containing the desired records
}
\author{
Thomas Goossens
}
