% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{makeDataset}
\alias{makeDataset}
\title{make a dataframe of stations records}
\usage{
makeDataset(user_token = Sys.getenv("AGROMET_API_V1_KEY"), stations,
  json = NULL, dfrom = NULL, dto = NULL, sensor = "tsa",
  staticExpl = "elevation", dynExpl = NULL)
}
\arguments{
\item{stations}{a character specifying the sid's of the stations to use separated by commas}

\item{json}{a character specifying the path of a json file constructed from a batch db export.
If NULL the API will be called.}

\item{dfrom}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{dto}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{sensor}{a character specifying the sensor data you want to spatialize.
One of tsa, hct, hra}

\item{staticExpl}{a character vector specifying the desired static explanatory variables
Any combinations of "altitude", "elevation", "slope", "aspect", "Agricultural_areas", "Artificials_surfaces", "Forest", "Herbaceous_vegetation". Latitude and longitude are always provided. Default = "Elevation"}

\item{dynExpl}{a character vector specifying the dynamic explanatory variables you want to add to the task.
Any combinations of inca, ens}

\item{token}{a character specifying your agrometAPI token}
}
\value{
a 2 elements named list : (1) snitch and (2) output. snitch is TRUE if function has provided the expected result. output is a named list which contains 3 elements :
(1) value : a list of dataframes where each dataframe contains the sid, the mtime, the sensor data, x position (longitude), y position (latitude) and the explanatory variables
(2) condition : a character specifying if the functions has encountered success, warning, error
(3) message : the message relative to the condition
}
\examples{
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T15:00:00Z",
  sensor = "tsa")

}
\author{
Thomas Goossens
}
