% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{makeDataset}
\alias{makeDataset}
\title{make a dataframe of stations records for each requested hour}
\usage{
makeDataset(user_token = Sys.getenv("AGROMET_API_V1_KEY"),
  api_request = "https://app.pameseb.be/agromet/api/v2/sp/cleandata",
  sensors = "tsa", stations = "all", stations_meta = stations.df,
  dfrom = NULL, dto = NULL, json = NULL, staticExpl = "elevation",
  dynExpl = NULL)
}
\arguments{
\item{sensors}{a character vector specifying the name of the sensors for which you data.
One of tsa, hct, hra, ens, plu, vvt.}

\item{stations}{a character vector specifying the sid's of the stations to use.}

\item{dfrom}{a datetime string specifying the dateTime from which you want data
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{dto}{a datetime string specifying the dateTime to which you want data
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{json}{a character specifying the path of a json file constructed from a batch db export.
If NULL the API will be called.}

\item{staticExpl}{a character vector specifying the desired static explanatory variables
Any combinations of "elevation", "slope", "aspect", "Agricultural_areas", "Artificials_surfaces", "Forest", "Herbaceous_vegetation". Latitude and longitude are always provided. Default = "Elevation"}

\item{dynExpl}{a character vector specifying the dynamic explanatory variables you want to add to the task.
Any combinations of inca, ens}

\item{token}{a character specifying your agrometAPI token}
}
\value{
A 2 elements named list
\itemize{
  \item \code{snitch} : a boolean. Is \code{TRUE} if function has provided the expected result. Is \code{FALSE} is function throws an error
  \item \code{output} : a named list which elements are :
    \itemize{
      \item \code{value} : a list which elements are of classes \code{data.frame} which colnames correspond to sid, mtime, the sensors, x (latitude), y (longitude) and the explanatory variables
      \item \code{condition} : a character specifying the condition encountered by the : success, warning, or error.
      \item \code{message} : a character specifying the message relative to the condition.
    }
 }
}
\examples{
\dontrun{
# get the dataset
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T15:00:00Z",
  sensor = "tsa")

# extract the data
myDataset = myDataset$output$value

# each dataframe is stored in a list
class(myDataset)

# show the head of the first dataframe
head(myDataset[[1]])
}

}
\author{
Thomas Goossens
}
