% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\title{make a mlr model + its metadata for a specific task using a specific learner}
\usage{
makeModel(task, learner)
}
\arguments{
\item{task}{an object of class mlr::makeRegrTask()}

\item{learner}{an object of class mlr::makeLearner()}
}
\value{
A 2 elements named list
\itemize{
  \item \code{snitch} : a boolean. Is \code{TRUE} if function has provided the expected result. Is \code{FALSE} is function throws an error
  \item \code{output} : a named list which elements are :\itemize{
    \item \code{value} : a named list which elements are : \itemize{
      \item \code{trained} : an element of class \code{mlr::makeWrappedMode()}
      \item \code{predictions} an element of class \code{data.frame} containing the predictions made by the trained model at station locations. Colnames are \code{truth} (= observations), \code{response} (= predictions), \code{se} (= standard error) and \code{residuals} (= truth - response)
      \item \code{perfs} : a named list which elements are : \itemize{
        \item \code{iters} : a dataframe containing the information relative to the leave-one-out cross validation of the model. Colnames are \code{learner.id}, \code{iter}, \code{rmse}, \code{mse}, \code{mae}, \code{timetrain}
        \item \code{agg} : a dataframe containing the aggregated performances of the LOOCV. Same colnames as \code{iters}
        }
     }
      \item \code{condition} : a character specifying the condition encountered by the function : success, warning, or error.
      \item \code{message} : a character specifying the message relative to the condition.
    }
 }
}
\description{
make a mlr model + its metadata for a specific task using a specific learner
}
\examples{
\dontrun{
# load magrittr for pipe use : \%>\%
library(magrittr)
# create the dataset
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T18:00:00Z",
  sensor = "tsa")

# extract the list of hourly sets of records
myDataset = myDataset$output$value

# create the tasks
myTasks = purrr::map(myDataset, makeTask, target = "tsa")

# extract the tasks from the outputs
myTasks = myTasks \%>\% purrr::modify_depth(1, ~.$"output"$"value"$"task")

# keep the first task
myTask = myTasks[[1]]

# create the model
myModel = makeModel(
  task = myTask,
  learner = agrometeorLearners$mulLR_lonLatAlt_NA)

# extract the relevant information
myModel = myModel$output$value

# show the results of the LOOCV of this hour forthe choosen learner
head(myModel$predictions)
head(myModel$perfs$iters)

# show the model information
myModel$trained
}
}
\author{
Thomas Goossens
}
