% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\title{make a mlr model + model metadata for a specific task using a specific learner}
\usage{
makeModel(task, learner)
}
\arguments{
\item{task}{an object of class mlr::makeRegrTask()}

\item{learner}{an object of class mlr::makeLearner()}
}
\value{
a list containing a boolean and an object of class mlr::train()
}
\author{
Thomas Goossens
}
