% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\title{make a mlr model + model metadata for a specific task using a specific learner}
\usage{
makeModel(task, learner)
}
\arguments{
\item{task}{an object of class mlr::makeRegrTask()}

\item{learner}{an object of class mlr::makeLearner()}
}
\value{
A 2 elements named list : \code{snitch} (2) output. \cr

A 2 elements named list : \code{snitch} & \code{output}. \cr
\code{output} is a named list which contains :
\itemize{
  \item \code{value} is an object of class list made of 3 elements which are of classes \code{mlr::makeWrappedMode()}, "dataframe", and "list"
  \item \code{condition} is a character specifying if the function has encountered success, warning, error.
  \item \code{message} is the message relative to the condition.
}
}
\examples{
\dontrun{
# create the dataset
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T18:00:00Z",
  sensor = "tsa")

# extract the list of hourly sets of records
myDataset = myDataset$output$value

# create the tasks
myTasks = purrr::map(myDataset, makeTask, target = "tsa")

# extract the tasks from the outputs
myTasks = myTasks \%>\% purrr::modify_depth(1, ~.$"output"$"value"$"task")

# keep the first task
myTask = myTasks[[1]]

# create the model
myModel = makeModel(
  task = myTask,
  learner = agrometeorLearners$mulLR_lonLatAlt_NA)

# extract the relevant information
myModel = myModel$output$value

# show the results of the LOOCV of this hour forthe choosen learner
head(myModel$predictions)
head(myModel$perfs$iters)

# show the model information
myModel$trained
}
}
\author{
Thomas Goossens
}
