% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{makePlots}
\alias{makePlots}
\title{make plots for a sensor of a station for a specific period of time}
\usage{
makePlots(records, sensor, sid, dfrom, dto, interactive = TRUE)
}
\arguments{
\item{sensor}{a character specifying the name of the sensor you want to plot}

\item{sid}{a character specifying the sid of the station you want to plot}

\item{dataset}{a dataframe containing the records from the 2 stations you want to compare. ::TODO ::Must contain sunset and sunrise !}
}
\value{
a list wihch elements are objects of class mlr::benchmark()
}
\author{
Thomas Goossens
}
