% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpatialization.R
\name{makeSpatialization}
\alias{makeSpatialization}
\title{spatialize a gridded dataset using a mlr model}
\usage{
makeSpatialization(model, pred.grid = grid.df)
}
\arguments{
\item{model}{an object of class mlr::train() that contains the prediction model}

\item{pred.grid}{an object of class sf::st_makegrid(). This object must contains the same column names as the task on which the model has been trained}
}
\value{
a 2 elements named list : (1) snitch and (2) output. snitch is TRUE if function has provided the expected result. output is a named list which contains
(1) value : an object of class data.frame containing the spatialized data
(2) condition : a character specifying if the functions has encountered success, warning, error
(3) message : the message relative to the condition
}
\examples{
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T15:00:00Z",
  sensor = "tsa")
myTask = makeTask(dataset = myDataset$output$value, target = "tsa")
myModel = makeModel(
  task = mytask$out$value,
  learner = learners$baseLearners$lrn.lm.alt)
mySpatialization = makeSpatialization(model = myModel$output$value)
}
\author{
Thomas Goossens
}
