% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpatialization.R
\name{makeSpatialization}
\alias{makeSpatialization}
\title{spatialize to geojson (EPSG = 4326) a gridded dataset using a mlr model}
\usage{
makeSpatialization(model, grid = grid.df, file = FALSE, filename = NULL,
  path = getwd())
}
\arguments{
\item{model}{an object of class mlr::train() that contains the prediction model}

\item{grid}{an object of class sf::st_makegrid(). This object must contains the same column names as the task on which the model has been trained}

\item{file}{a boolean specifying wether the saptiliazed dat should be written as a geojson file or simply output into the console.}

\item{filename}{a character specifying the name you want to give to the file.}

\item{path}{a character specifying the path where you want your geosonfile to be stored. Default = wd}
}
\value{
a geojson object containing the spatialized data
}
\author{
Thomas Goossens
}
