% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpatialization.R
\name{makeSpatialization}
\alias{makeSpatialization}
\title{spatialize a gridded dataset using a mlr model}
\usage{
makeSpatialization(model, pred.grid = grid.df)
}
\arguments{
\item{model}{an object of class mlr::train() that contains the prediction model}

\item{pred.grid}{an object of class sf::st_makegrid(). This object must contains the same column names as the task on which the model has been trained}
}
\value{
A 2 elements named list : \code{snitch} & \code{output}. \cr
\code{snitch} is \code{TRUE} if function has provided the expected result. \cr
\code{output} is a named list which contains :
\itemize{
  \item \code{value} is an object of class list where first element is a dataframe containing the spatialized data and the second element is a data.frame containing the summary data
  \item \code{condition} is a character specifying if the function has encountered success, warning, error.
  \item \code{message} is the message relative to the condition.
}
}
\examples{
\dontrun{
# create the dataset
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T18:00:00Z",
  sensor = "tsa")

# extract the list of hourly sets of records
myDataset = myDataset$output$value

# create the tasks
myTasks = purrr::map(myDataset, makeTask, target = "tsa")

# extract the tasks from the outputs
myTasks = myTasks \%>\% purrr::modify_depth(1, ~.$"output"$"value"$"task")

# keep the first task
myTask = myTasks[[1]]

# create the model
myModel = makeModel(
  task = myTask,
  learner = agrometeorLearners$mulLR_lonLatAlt_NA)

# extract the relevant information
myModel = myModel$output$value

# spatialize using the trained model
mySpatialization = makeSpatialization(
model = myModel$trained,
pred.grid = grid.df) # grid.df comes precompiled with the package

# get the relevant information
mySpatialization = mySpatialization$output$value

# show an excerpt of the spatialized data
head(mySpatialization$spatialized)

# show the summary stats of spatialized data
head(mySpatialization$summary)
}
}
\author{
Thomas Goossens
}
