% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTask.R
\name{makeTask}
\alias{makeTask}
\title{make a single mlr regression task for a single set of hourly/daily records}
\usage{
makeTask(dataset, drop = NULL, target)
}
\arguments{
\item{dataset}{a dataframe containing an hourly/daily set of records you want to transform to a mlr task}

\item{drop}{a character vector specifying the explanatory variables you want to drop.}

\item{target}{a charachter specifying the name of the target variable}
}
\value{
A 2 elements named list
\itemize{
  \item \code{snitch} : a boolean. Is \code{TRUE} if function has provided the expected result. Is \code{FALSE} is function throws an error
  \item \code{output} : a named list which elements are :
    \itemize{
      \item \code{value} : a list which elements are of classes \code{mlr::RegrTask()}
      \item \code{condition} : a character specifying the condition encountered by the : success, warning, or error.
      \item \code{message} : a character specifying the message relative to the condition.
      \item \code{stations} : a named list which elements are  \itemize{
        \item \code{used} : an integer vector of the sids of the used stations for the task creation
        \item \code{ignored} an integer vector of the sids of the ignored stations for the task creation
        }
    }
 }
}
\examples{
\dontrun{
# create the dataset
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T18:00:00Z",
  sensor = "tsa")

# extract the list of hourly sets of records
myDataset = myDataset$output$value

# create the tasks
myTasks = purrr::map(myDataset, makeTask, target = "tsa")

# extract the tasks from the outputs
myTasks = myTasks \%>\% purrr::modify_depth(1, ~.$"output"$"value"$"task")

# show the first task information
myTasks[[1]]

# extract the lists of stations kept for each task
stations_kept = myTasks \%>\% purrr::modify_depth(1, ~.$"output"$"value"$"stations")

# show the stations kept for the first task
stations_kept[[1]]
}
}
\author{
Thomas Goossens
}
