% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTask.R
\name{makeTask}
\alias{makeTask}
\title{make one mlr regr task for a single set of hourly/daily records}
\usage{
makeTask(dataset, drop = NULL, target)
}
\arguments{
\item{dataset}{a dataframe containing an hourly/daily set of records you want to transform to a mlr task}

\item{drop}{a character vector specifying the explanatory variables you wan to drop.}

\item{target}{a charachter specifying the name of the target variable}
}
\value{
a list containing a boolean and a list which single element is of class mlr::makeRegrTask()
}
\author{
Thomas Goossens
}
