% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTasks.R
\name{makeTasks}
\alias{makeTasks}
\title{make a mlr regr task for an hourly/daily set of records}
\usage{
makeTasks(stations = paste0(as.character(stations.df$sid), collapse = ","),
  dfrom, dto, sensor = "tsa", dynExpl = NULL, drop = c("sid", "altitude",
  "slope", "aspect", "Agricultural_areas", "Artificials_surfaces", "Forest",
  "Herbaceous_vegetation"))
}
\arguments{
\item{stations}{a character specifying the sid's of the stations to use separated by commas}

\item{dfrom}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{dto}{a datetime string specifying the dateTime
Must have the form "YYYY-MM-DDTHH:MM:SSZ"}

\item{sensor}{a character specifying the sensor data you want to spatialize.
One of tsa, hct, hra}

\item{dynExpl}{a character vector specifying the dynamic explanatory variables you want to add.
Any combinations of inca, ens}

\item{drop}{a character vector specifying the explanattory variables you want to drop.
Any combinations of "altitude", "elevation", "slope", "aspect", "Agricultural_areas", "Artificials_surfaces", "Forest","Herbaceous_vegetation"}
}
\value{
an object of class list containing an mlr::makeRegrTask() and a sumamry dataframe of the task dataset
}
\author{
Thomas Goossens
}
