% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTasks.R
\name{makeTasks}
\alias{makeTasks}
\title{make one or multiple mlr regr task(s) for sets of hourly/daily records}
\usage{
makeTasks(dataset, drop = NULL, target)
}
\arguments{
\item{dataset}{a dataframe containing all the hourly/daily records you want to transform to a list of mlr tasks}

\item{drop}{a character vector specifying the explanatory variables you wan to drop.}

\item{target}{a charchter specifying the name of the target variable}

\item{stations}{a character specifying the sid's of the stations to use separated by commas}
}
\value{
a list containing a boolean and a list which elements are of class mlr::makeRegrTask()
}
\author{
Thomas Goossens
}
